backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
supabase==2.3.0
pydantic==2.5.3
numpy==1.26.3
scikit-learn==1.4.0
tensorflow-macos==2.15.0; sys_platform == 'darwin' and platform_machine == 'arm64'
tensorflow-metal==1.1.0; sys_platform == 'darwin' and platform_machine == 'arm64'
tensorflow==2.15.0; sys_platform != 'darwin' or platform_machine != 'arm64'
requests==2.31.0
